requirements.txt
python-telegram-bot==12.8
google_trans_new
requests
